# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 63307f5ac97c70218d65d9e0
  title: Zuct
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /user:
    post:
      x-skaffolder-id: 63307f5bc97c70218d65d9f2
      x-skaffolder-name: create
      x-skaffolder-id-resource: 63307f5ac97c70218d65d9ea
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 63307f5bc97c70218d65d9ed
      x-skaffolder-name: list
      x-skaffolder-id-resource: 63307f5ac97c70218d65d9ea
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 63307f5bc97c70218d65d9ee
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 63307f5ac97c70218d65d9ea
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 63307f5bc97c70218d65d9f3
      x-skaffolder-name: get
      x-skaffolder-id-resource: 63307f5ac97c70218d65d9ea
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 63307f5bc97c70218d65d9f0
      x-skaffolder-name: update
      x-skaffolder-id-resource: 63307f5ac97c70218d65d9ea
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
...
  /students:
    post:
      x-skaffolder-id: 63308790c97c70218d65da8a
      x-skaffolder-name: create
      x-skaffolder-id-resource: 63308322c97c70218d65da4e
      x-skaffolder-resource: students
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - students
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/students"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date_of_birth:
                  type: integer
                first_name:
                  type: string
                last_name:
                  type: string
                middle_name:
                  type: string
                phone:
                  type: number
                sex:
                  type: string
    get:
      x-skaffolder-id: 63308790c97c70218d65da8f
      x-skaffolder-name: list
      x-skaffolder-id-resource: 63308322c97c70218d65da4e
      x-skaffolder-resource: students
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - students
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/students"
  "/students/{id}":
    delete:
      x-skaffolder-id: 63308790c97c70218d65da8e
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 63308322c97c70218d65da4e
      x-skaffolder-resource: students
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - students
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 63308790c97c70218d65da8b
      x-skaffolder-name: get
      x-skaffolder-id-resource: 63308322c97c70218d65da4e
      x-skaffolder-resource: students
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - students
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/students"
    post:
      x-skaffolder-id: 63308790c97c70218d65da88
      x-skaffolder-name: update
      x-skaffolder-id-resource: 63308322c97c70218d65da4e
      x-skaffolder-resource: students
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - students
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/students"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/students"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      x-skaffolder-id: 63307f5ac97c70218d65d9ea
      x-skaffolder-id-db: 63307f5ac97c70218d65d9e1
      x-skaffolder-id-entity: 63307f5ac97c70218d65d9e2
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 63307f5ac97c70218d65d9e5
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 63307f5ac97c70218d65d9e6
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 63307f5ac97c70218d65d9e4
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 63307f5ac97c70218d65d9e8
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 63307f5ac97c70218d65d9e7
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 63307f5ac97c70218d65d9e3
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...
    students:
      x-skaffolder-id: 63308322c97c70218d65da4e
      x-skaffolder-id-db: 63307f5ac97c70218d65d9e1
      x-skaffolder-id-entity: 63308322c97c70218d65da47
      x-skaffolder-url: /students
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        date_of_birth:
          type: integer
          x-skaffolder-id-attr: 63308322c97c70218d65da4c
          x-skaffolder-type: Date
          x-skaffolder-required: true
        first_name:
          type: string
          x-skaffolder-id-attr: 63308322c97c70218d65da48
          x-skaffolder-type: String
          x-skaffolder-required: true
        last_name:
          type: string
          x-skaffolder-id-attr: 63308322c97c70218d65da4a
          x-skaffolder-type: String
          x-skaffolder-required: true
        middle_name:
          type: string
          x-skaffolder-id-attr: 63308322c97c70218d65da49
          x-skaffolder-type: String
        phone:
          type: number
          x-skaffolder-id-attr: 63308322c97c70218d65da4d
          x-skaffolder-type: Number
        sex:
          type: string
          x-skaffolder-id-attr: 63308322c97c70218d65da4b
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations:
        _enrolls:
          x-skaffolder-id: 63308a42c97c70218d65da94
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 63308322c97c70218d65da47
          x-skaffolder-ent2: 633086dfc97c70218d65da64
        _takes:
          x-skaffolder-id: 63308a42c97c70218d65da93
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 63308322c97c70218d65da47
          x-skaffolder-ent2: 63308546c97c70218d65da51
        _taught:
          x-skaffolder-id: 63308a42c97c70218d65da95
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 63308322c97c70218d65da47
          x-skaffolder-ent2: 633085b7c97c70218d65da59
      required:
        - _id
        - date_of_birth
        - first_name
        - last_name
        - sex
...